requirements.txt

PyQt5==5.15.9
requests==2.31.0
dnspython==2.4.2
python-whois==0.8.0
beautifulsoup4==4.12.2
pycryptodome==3.18.0
reportlab==4.0.4
jinja2==3.1.2
python-nmap==0.7.1
aiohttp==3.9.1
asyncio==3.4.3
aiodns==3.1.1
urllib3==2.0.7
